requirements.txt
alembic==1.13.2
babel==2.16.0
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
packaging==24.1
psycopg2==2.9.9
python-dotenv==1.0.1
speaklater==1.3
SQLAlchemy==2.0.34
typing_extensions==4.12.2
Werkzeug==3.0.4
